ETL/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pyodbc==5.2.0
pandas==2.2.3
requests==2.32.3
tensorflow==2.18.0
numpy==2.0.2